refactor: getUser 함수 data.ts 로 이동
