+++ app/lib/data.ts
@@ -1,4 +1,5 @@
 import db from '@/lib/db';
+import getSession from '@/lib/session';
 
 export async function fetchBooks(userId: number) {
   try {
@@ -19,5 +20,22 @@ export async function fetchBooks(userId: number) {
   } catch (error) {
     console.error('Database Error:', error);
     throw new Error(`Failed to fetch books of user(id : ${userId})`);
+  }
+}
+
+export async function getUser() {
+  try {
+    const session = await getSession();
+    if (session.id) {
+      const user = await db.user.findUnique({
+        where: {
+          id: session.id,
+        },
+      });
+      if (user) return user;
+    }
+  } catch (error) {
+    console.error('Failed to fetch user:', error);
+    throw new Error('Failed to fetch user.');
   }
 }
